nvme: add 'nr_sqs' and 'nr_cqs' to nvme_ctrl
Some of public APIs in nvme/queue takes 'struct nvme_sq' and 'struct
nvme_cq' pointer as an argument.  To retrieve a specific queue instance
from ctrl->sq or ctrl->cq, caller should know the range of the array
which is dynamically allocated in nvme_ctrl_init().

This patch added 'nr_sqs' and 'nr_cqs' members to struct nvme_ctrl so
that caller exactly knows the limit of the array range to access.

Signed-off-by: Minwoo Im <minwoo.im.dev@gmail.com>